+++ cmd/login.go
@@ -27,22 +27,3 @@ func newLoginCmd() *cobra.Command {
 	cmd.PersistentFlags().StringVarP(&cloudURL, "cloud-url", "c", "", "A cloud URL to log into")
 	return cmd
 }
-
-func newLogoutCmd() *cobra.Command {
-	var cloudURL string
-	cmd := &cobra.Command{
-		Use:   "logout",
-		Short: "Log out of the Pulumi Cloud",
-		Long:  "Log out of the Pulumi Cloud.  Deletes stored credentials on the local machine.",
-		Args:  cmdutil.NoArgs,
-		Run: cmdutil.RunFunc(func(cmd *cobra.Command, args []string) error {
-			if cloudURL == "" {
-				// If no URL was specified, assume it's the default URL.
-				cloudURL = cloud.DefaultURL()
-			}
-			return cloud.Logout(cloudURL)
-		}),
-	}
-	cmd.PersistentFlags().StringVarP(&cloudURL, "cloud-url", "c", "", "A cloud URL to log out of")
-	return cmd
-}
